+++ src/core/lombok/javac/handlers/singulars/JavacJavaUtilSetSingularizer.java
@@ -40,6 +40,12 @@ public class JavacJavaUtilSetSingularizer extends JavacJavaUtilListSetSingulariz
 		return LombokImmutableList.of("java.util.Set", "java.util.SortedSet", "java.util.NavigableSet");
 	}
 	
+	@Override protected String getEmptyMaker(String target) {
+		if (target.endsWith("SortedSet")) return "java.util.Collections.emptySortedSet";
+		if (target.endsWith("NavigableSet")) return "java.util.Collections.emptyNavigableSet";
+		return "java.util.Collections.emptySet";
+	}
+	
 	@Override public void appendBuildCode(SingularData data, JavacNode builderType, JCTree source, ListBuffer<JCStatement> statements, Name targetVariableName, String builderVariable) {
 		JavacTreeMaker maker = builderType.getTreeMaker();