+++ test/transform/resource/before/SuperBuilderWithGenericsAndToBuilder.java
@@ -1,10 +1,10 @@
-import java.util.List;
+import java.util.Map;
 
 public class SuperBuilderWithGenericsAndToBuilder {
 	@lombok.experimental.SuperBuilder(toBuilder = true)
 	public static class Parent<A> {
 		A field1;
-		@lombok.Singular List<String> items;
+		@lombok.Singular Map<Integer, String> items;
 	}
 	
 	@lombok.experimental.SuperBuilder(toBuilder = true)
@@ -13,6 +13,6 @@ public class SuperBuilderWithGenericsAndToBuilder {
 	}
 	
 	public static void test() {
-		Child<Integer> x = Child.<Integer>builder().field3(0.0).field1(5).item("").build().toBuilder().build();
+		Child<Integer> x = Child.<Integer>builder().field3(0.0).field1(5).item(5, "").build().toBuilder().build();
 	}
 }